Add support for multipath devices
Previously, when a device-mapper multipath device was specified for DEVS
setup would fail with "Partition specification unsupported at this
time." Each device is checked to see if it has a parent, and, if so, it
is assumed to be a partition.  Multipath devices are not partitions, and
their backing paths are reported as parent devices.

This change adds the ability to correctly identify multipath devices by
checking for device type of "mpath" so that they are not mistaken for
partitions.